FlaskApp/requirements.txt
Flask==0.12.2
Jinja2==2.10
python-firebase==1.2
